+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 #### Version 0.1.2 June 2015
 - **(A)** Added WhatsApp to sorting field.
 - **(A)** Integrated CoalaWeb News into version field.
+- **(C)** CSS improvements
 - **(C)** Updated CoalaWeb's custom prefix version of Uikit to 2.21.
+- **(B)** Fix bug when adding new items to sorting field.
 
 #### Version 0.1.1 March 2015
 - **(A)** New popup close HTML file.